Fully implemented review system
When on homepage allows user to create a review if they are logged in and not created a review that day. Stores in review table. Retrieves table and prints it whenever home screen is viewed.
Also redirects the base url (no /) to the /index url to avoid having to manually go to /index